+++ test/test_basis_consistency.jl
@@ -1,5 +1,6 @@
 using Test
 using QuantumSymbolics
+using QuantumOptics
 
 @test express(Z*Z1) == express(Z1)
 @test express(Z*Z2) == -express(Z2)
@@ -11,3 +12,8 @@ using QuantumSymbolics
 @test express(Pp*Z2) == express(Z1)
 @test express(Pm*L0) == express(L1)
 @test express(Pp*L1) == express(L0)
+
+@op A; @op B; @op C; @op O; @ket k;
+@superop S; K = kraus(A, B, C);
+
+@test basis(K) == basis(A)